+++ lunes_cms/__init__.py
@@ -1,4 +1,4 @@
 """
 Content Management System for the Lunes Vocabulary Trainer App
 """
-__version__ = "2022.6.0"
+__version__ = "2022.6.1"