+++ CHANGELOG.md
@@ -2,6 +2,7 @@ UNRELEASED
 ----------
 
 * [ [#337](https://github.com/digitalfabrik/lunes-cms/issues/337) ] Remove image names
+* [ [#337](https://github.com/digitalfabrik/lunes-cms/issues/211) ] Add feedback endpoint to API
 
 
 2022.6.0