+++ CHANGELOG.md
@@ -4,6 +4,7 @@ UNRELEASED
 * [ [#337](https://github.com/digitalfabrik/lunes-cms/issues/337) ] Remove image names
 * [ [#211](https://github.com/digitalfabrik/lunes-cms/issues/211) ] Add feedback endpoint to API
 * [ [#346](https://github.com/digitalfabrik/lunes-cms/issues/346) ] Add bulk action to mark feedback as read
+* [ [#317](https://github.com/digitalfabrik/lunes-cms/issues/317) ] Add icon preview for disciplines and training sets
 
 
 2022.6.0