+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 UNRELEASED
 ----------
 
+
+2022.6.1
+--------
+
 * [ [#337](https://github.com/digitalfabrik/lunes-cms/issues/337) ] Remove image names
 * [ [#211](https://github.com/digitalfabrik/lunes-cms/issues/211) ] Add feedback endpoint to API
 * [ [#346](https://github.com/digitalfabrik/lunes-cms/issues/346) ] Add bulk action to mark feedback as read